supbase/backend/requirements.txt
Flask
requests
carbon-python-sdk
supabase